backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pgvector==0.3.6

# Redis
redis[hiredis]==5.2.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# AI / LLM
anthropic==0.43.0
litellm==1.55.8

# Background tasks
apscheduler==3.10.4

# HTTP client (for agent MC client)
httpx==0.28.1

# Utilities
python-dotenv==1.0.1
orjson==3.10.13

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# Linting
ruff==0.8.6
